+++ packages/arcgis-rest-feature-service/test/mocks/feature.ts
@@ -1,0 +1,13 @@
+import { IFeature } from "../../src/features";
+
+export const featureResponse = {
+  feature: {
+    attributes: {
+      FID: 42
+    },
+    geometry: {
+      x: -9177311.62541634,
+      y: 4247151.205222242
+    }
+  }
+};